--- packages/design-system/src/components/button/Button.tsx
@@ -1,25 +1,0 @@
-import { ComponentProps, ReactNode } from "react";
-import { cn } from "../../lib/cn";
-
-interface ButtonProps extends ComponentProps<"button"> {
-  size?: "small" | "large";
-}
-
-export default function Button({
-  size = "small",
-  disabled = false,
-  children,
-  ...rest
-}: ButtonProps) {
-  const baseStyle =
-    "rounded-md font-medium bg-pink-200 transition-colors duration-200 disabled:cursor-not-allowed disabled:bg-gray-400 disabled:text-white";
-
-  const sizeStyle =
-    size === "large" ? "text-lg px-6 py-3" : "text-sm px-4 py-2";
-
-  return (
-    <button className={cn(baseStyle, sizeStyle)} disabled={disabled} {...rest}>
-      {children}
-    </button>
-  );
-}
